requirements.txt
streamlit>=1.28.0
Pillow>=9.5.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.15.0
cryptography>=41.0.0
mutagen>=1.47.0
reportlab>=4.0.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
pandas>=2.0.0
